requirements.txt
adal==1.2.2
alembic==1.4.1
audioread==2.1.8
azure-datalake-store==0.0.48
Bootstrap-Flask==1.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
dominate==2.4.0
Flask==1.1.1
Flask-Admin==1.5.5
flask-behind-proxy==0.1.1
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.0.3
numba==0.47.0
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
psycopg2==2.8.4
pycparser==2.19
pydub==0.23.1
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.13
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
